Eliminando

+++ src/app/lista-comics/lista-comics.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ComicServiceService } from '../comic-service.service';
 import { AuthService } from '../auth.service';
 import { ComicInsertar } from '../model/comicInsertar';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-lista-comics',
@@ -29,16 +30,48 @@ export class ListaComicsComponent implements OnInit {
     console.log(comic);
     this.comicService.actualizarComic(comic).subscribe(data => {
       console.log(data);
+
+      Swal.fire({
+        type: "success",
+        title: "Actualización correcta",
+        text: "Se ha actualizado el comic"
+      })
+
     });
   }
 
   onDelete(comic: ComicInsertar) {
-    console.log("Borrar" + comic.idComics);
-    this.comicService.eliminarComic(comic).subscribe(data => {
-      console.log(data);
-      this.listarComics();
-    });
-  }
+
+
+    Swal.fire({
+      title:'¿Eliminar Comic?',
+      text: "Esta acción no se puede revertir",
+      type:'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+    }).then((result) =>{
+      if(result.value){
+        console.log("Borrar" + comic.idComics);
+        this.comicService.eliminarComic(comic).subscribe(data => {
+          console.log(data);
+          this.listarComics();
+
+            Swal.fire({
+              type: "success",
+              title: "Eliminado correctamente",
+              text: "Se ha eliminado al usuario del sistema"
+            }).then(function(){
+              location.reload();
+            })
+      })
+    }
+
+
+  })
+
+}
 
 
 }
